+++ migration-library/src/main/kotlin/com/quadient/migration/api/dto/migrationmodel/documentcontent/Paragraph.kt
@@ -59,6 +59,7 @@ data class Paragraph(
                     is FirstMatch -> textContent.toDb()
                     is Hyperlink -> textContent.toDb()
                     is ColumnLayout -> textContent.toDb()
+                    is Barcode -> textContent.toDb()
                 }
             }.toMutableList(), it.styleRef?.toDb(), it.displayRuleRef?.toDb())
         }.toMutableList(),