+++ migration-library/src/main/kotlin/com/quadient/migration/service/inspirebuilder/DesignerDocumentObjectBuilder.kt
@@ -7,6 +7,7 @@ import com.quadient.migration.api.repository.DocumentObjectRepository
 import com.quadient.migration.api.repository.ParagraphStyleRepository
 import com.quadient.migration.api.repository.Repository
 import com.quadient.migration.api.repository.TextStyleRepository
+import com.quadient.migration.api.repository.VariableRepository
 import com.quadient.migration.service.imageExtension
 import com.quadient.migration.service.ipsclient.IpsService
 import com.quadient.migration.service.resolveAliases
@@ -48,7 +49,7 @@ class DesignerDocumentObjectBuilder(
     documentObjectRepository: DocumentObjectRepository,
     textStyleRepository: TextStyleRepository,
     paragraphStyleRepository: ParagraphStyleRepository,
-    variableRepository: Repository<Variable>,
+    variableRepository: VariableRepository,
     variableStructureRepository: Repository<VariableStructure>,
     displayRuleRepository: Repository<DisplayRule>,
     imageRepository: Repository<Image>,
@@ -295,7 +296,17 @@ class DesignerDocumentObjectBuilder(
         languages: List<String>
     ): Flow? {
         val flow = getFlowByName(layout, documentModel.nameOrId())
-            ?: if (documentModel.internal == true || documentModel.type == DocumentObjectType.Snippet) {
+            ?: if (documentModel.type == DocumentObjectType.Snippet) {
+                buildDocumentContentAsSingleFlow(
+                    layout,
+                    variableStructure,
+                    documentModel.content,
+                    documentModel.nameOrId(),
+                    documentModel.displayRuleRef?.let { DisplayRuleRef(it.id) },
+                    languages,
+                    isInline = true
+                )
+            } else if (documentModel.internal == true) {
                 buildDocumentContentAsSingleFlow(
                     layout,
                     variableStructure,
@@ -338,6 +349,7 @@ class DesignerDocumentObjectBuilder(
             when (it) {
                 is Area -> pageContentModels.add(PageContent.AreaContent(it))
                 is Shape -> pageContentModels.add(PageContent.PathObjectContent(it))
+                is Barcode -> pageContentModels.add(PageContent.BarCodeContent(it))
                 else -> virtualAreaContent.add(it)
             }
         }
@@ -349,6 +361,9 @@ class DesignerDocumentObjectBuilder(
                 }
                 is PageContent.PathObjectContent -> {
                     buildPathObject(layout, page, model.content)
+                }
+                is PageContent.BarCodeContent -> {
+                    model.content.buildContent(variableRepository, page.barcodeFactory, layout, variableStructure, false)
                 }
             }
         }
@@ -503,5 +518,6 @@ class DesignerDocumentObjectBuilder(
     private sealed interface PageContent {
         data class AreaContent(val content: Area): PageContent
         data class PathObjectContent(val content: Shape): PageContent
+        data class BarCodeContent(val content: Barcode): PageContent
     }
 }