+++ wfd-xml/api/src/main/java/com/quadient/wfdxml/api/layoutnodes/BarcodeFactory.java
@@ -3,4 +3,6 @@ package com.quadient.wfdxml.api.layoutnodes;
 public interface BarcodeFactory {
 
     DataMatrixBarcode addDataMatrix();
+    QrBarcode addQr();
+    Code39Barcode addCode39();
 }