+++ wfd-xml/api/src/main/java/com/quadient/wfdxml/api/layoutnodes/QrBarcode.java
@@ -1,0 +1,13 @@
+package com.quadient.wfdxml.api.layoutnodes;
+
+public interface QrBarcode extends Barcode<QrBarcode> {
+    QrBarcode setModuleWidth(double moduleWidth);
+
+    QrBarcode setWhiteSpace(double whiteSpace);
+
+    QrBarcode setBarcodeSize(int sizeEnum);
+
+    QrBarcode setEciCode(int eciCodeEnum);
+
+    QrBarcode setErrorCorrection(int errorCorrectionEnum);
+}